% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuctions.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string function}
\usage{
str_split_one(string, split)
}
\arguments{
\item{string}{A character vector with one element}

\item{split}{What to split on}
}
\value{
A character vector
}
\description{
Split a string function
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(string = x, split = ",")
}
